Separate KStream reader from writer
For simplicity and modularity the reader and writer KStream is become
separated. This commit removes all writer parts from the KStream class.
Also, the mode is brought to the type instead of being member variable.
This keeps the design simple and allows us to write specialised methods
for each of them separately.