Add double-buffer writer
Add the KStream writer class with double buffer writing strategy. It is
multi-threaded, one thread watching the second buffer and write it down
to the file whenever is updated. The main thread write to the first
buffer instead of actual writing to the file and whenever the first
buffer is full, it copies the content to the second one and notify the
writer.