Add a mechanism to force a specific output format
The format of a record is determined by its quality string; i.e. if a record has
a non-empty quality string, it is considered as a FASTQ record. Otherwise, it is
written in FASTA format. This commit fixes #1 by adding some modifiers to force
a specific file format. The format can be set either by `set_format` method or
by passing the format using `operator<<`; e.g. to force a record in FASTQ
format:

    out << format::fastq << record;