Change default seq/qual line wrapping for FASTQ
Since most downstream tools does not support multiline squence (or
quality string) in FASTQ format, it would be sensible to output
single-line records by default (in response to issue #6). This only
applies when the format is explicitly set to `format::fastq`.

`SeqStreamOut` now accepts the output format and forward it to the base
class `KStreamOut`.

The line wrapping of `seq` and `qual` strings can be disabled by setting
the `wraplen` to zero. This commit also fixes the bug when the `wraplen`
is set to zero causing the program stuck in an infinite loop.

The `KStreamOut::puts` method no longer get `wrap` parameter as wrapping
is determined by the value of `KStreamOut::wraplen`.